requirements.txt
annotated-types
anyio
certifi
charset-normalizer
click
cmake
face-recognition
face_recognition_models
fastapi
idna
numpy
opencv-python
pillow
pydantic
pydantic_core
requests
setuptools
sniffio
starlette
typing_extensions
urllib3
uvicorn